ok to set uv to a fraction
Ignore-this: 9f328983eb589c131730a2fc21fa61e6

darcs-hash:20150613220136-312f9-417ff75c5ee83333ce493821e3ff81c37600c41b

+++ light9/effecteval/effectloop.py
@@ -208,7 +208,7 @@ class ControlBoard(object):
         level: 0 to 1
         """
         command = {0: '\x02', 1: '\x03'}[which]
-        self._dev.write('\x60' + command + chr(max(0, min(1, level)) * 255))
+        self._dev.write('\x60' + command + chr(int(max(0, min(1, level)) * 255)))
         self._dev.flush()
 
     def setRgb(self, color):
